% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpa.R
\name{vpa}
\alias{vpa}
\title{Virtual Population Analysis}
\usage{
vpa(C, M, Fterm, Fages)
}
\arguments{
\item{C}{catch-at-age matrix.}

\item{M}{natural mortality rate, a scalar.}

\item{Fterm}{terminal F, either a scalar or a vector of same length as the
number of ages.}

\item{Fages}{number of ages to calculate F for oldest age. For example, if
the catch-at-age matrix contains ages up to 15 and \code{Fages=5},
then F for the oldest age will be set as the average of ages 10-14.}
}
\value{
List containing the model results (\code{N}, \code{F}, \code{Z}), as well as
the input objects passed by the user (\code{C}, \code{M}, \code{Fterm},
\code{Fages}).
}
\description{
Run simple VPA model, with no tuning index.
}
